requirements.txt
streamlit==1.32.2
langchain==0.1.13
langchain-community==0.0.29
langchain-groq==0.0.1
chromadb==0.4.24
sentence-transformers==2.6.0
python-dotenv==1.0.1
duckduckgo-search==5.2.1
pypdf==4.1.0
